commitment+proof: move asset proof encoders/records
To avoid yet another circular dependency, we move the commitment proof
related encoders and records into the commitment package.
